Adapt the test suite to use the logging system properly
Given the previous changes to warning and error reporting, the test
suite needs new helper code to deal with them.  Again, the changes to
mosts tests are straight-forward (the location and formatting of
messages changes slightly), but the new logging machinery itself now
also needs to be tested... particularly the code that hooks into
Python's logging system, into Python's warning system, and between those
two.

`pytest` actually includes test fixtures to capture and interact with
warnings and logging calls, but these work on the library level:
warnings are captured as warnings, logging calls are captured as log
records.  To be useful to application testing, it is necessary to
inspect and assert both the low-level warnings/log records as well as
the standard error output they ultimately translate to.  The
`hypothesis`-based tests need particular (manual) care, because
`hypothesis` rejects tests using function-scoped `pytest` fixtures.