Explicitly support Pageant 0.82 in the test suite
Pageant 0.82, released two weeks ago, supports both proper output buffer
flushing and a "foreground" mode.  Foreground mode is used to correctly
isolate a spawned Pageant instance in the test suite.  The test for
Pageant's suitability can now be simplified to whether it is at least
version 0.82, or not.

Furthermore, because no published version of Pageant supports only
proper output buffer flushing but not foreground mode, it is no longer
necessary to arrange for debug output to be silently discarded.  This
further simplifies SSH agent handling considerably.

Finally, we fix various mistakes in the documentation of the testing
machinery.