Allow the user to overwrite the current vault configuration
In the "vault" derivation scheme, commands that edit or import
a configuration actually merge with the existing configuration, for
compatibility with the original vault(1).  However, sometimes it is
desired to replace the existing configuration instead, when importing
or when setting a (service or global) configuration.  We introduce a new
`--overwrite-existing` command-line option that signals exactly that,
and a `--merge-existing` option which signals the (previously implicit)
default merging behavior.