Fix usage of `--debug`, `--verbose` and `--quiet` options
The three "verbosity" options `--debug`, `--verbose` and `--quiet` set
the logging level for `derivepassphrase`, globally.  Because `click`
does not support mutually exclusive options by default, these options
are implemented as boolean flags, and it is possible to override earlier
verbosity settings with later options.

The previous setup treated the three options as disjointed flags, so
each callback was called once, irrespective of the others, so the
aforementioned intended overriding behavior was not working properly.
In addition, the callback was faulty, always enabling each of the flags.

The new setup now links all three options to the same flag, with
different flag values.  We further expose the option objects at the top
level, and unify the three callback functions to a single one.

As a side effect, we also fix a misclassified diagnostic during
configuration file migration ("vault" derivation scheme).