Make obtaining the compatibility config filename more explicit
When obtaining the compatibility config filename `settings.json` from
`cli._config_filename`, rename the corresponding subsystem from
`settings` to `old settings.json`.  Furthermore, pass that explicit
subsystem upon each call.  This frees us up to change the default
subsystem in a future commit.

Because this is not a public function, this is not a change in the
public API.

Some reformatting is necessary.