Make the mutually exclusive options helper function print the correct parameter name
The `click.Parameter.human_readable_name` attribute yields the metavar
for arguments, and the function parameter name for options.  The
documentation calls the latter the "name" for the option, perhaps under
the assumption that you would never pass an explicit parameter name to
the option decorator.  I disagree with this, it is horrible design:
sometimes the value is usable as a function parameter (options),
sometimes it isn't (arguments), and vice versa for constructing
a command-line, so it is neither usable as a command-line parameter
display name nor as a function parameter name.

So what *is* it even good for...?

Include our own logic to select an appropriate display name for our
options.  This requires querying `click.Parameter` attributes that
should be public, but are internal.  Why...?!