Support suppressing or forcing color output
We implement this with an eager, hidden "pseudo" option that checks for
the `NO_COLOR` and `FORCE_COLOR` options and sets the context's color
setting appropriately.

To actually make use of this color setting, all output must query the
context's color setting, as must all log output.  For the log output
specifically, we arrange to pass the `click` context in the extra dict
of each logging call, so that the handler has access to the color
setting too.