Fix check of empty service name for `derivepassphrase vault`
We asserted in c1bf00eadd1bf733ac25a25eafbe110d61936c54 that an empty
service name is both diagnosed, then treated as if it were missing later
on.  This wasn't actually strictly true: the checks for a missing or
superfluous service name tested for truthy service names instead of
non-missing service names, and the key selection and passphrase entry
dialogs happen before the check for a non-empty service name.  Fix these
two issues by checking for non-missing service names in the former case,
and moving the non-empty service name check further to the front in the
latter case.