Consolidate shell completion, add tests and fix Zsh output
We move the shell completion code in `derivepassphrase/cli.py` to a new
section and retire the specific handling of `VAULT_PATH` for the two
path completion functions.  Having actually tried this out interactively
in Bash, it does not work that well if we complete filenames, partial
filenames, and fixed strings in the same completion function: you would
have to set conflicting completion options for this function.  It *also*
does not work at all with `click`'s stock Bash completion script, which
would then discard either the fixed strings and partial filenames or the
complete filenames, depending on the order they are emitted in by our
shell completion code.  So abandon the completion handling of
`VAULT_PATH` in favor of standard filename completion, for paths.

Every supported shell has further limitations on which inputs it can
properly deserialize: Bash strips NUL characters from command
substitutions, and Fish (v3) breaks in most situations involving
newlines.  Furthermore, the stock shell completion scripts add their own
additional limitations: all shells use newline-terminated messages, so
embedded newlines in the completion item (type, value, help text) cause
(generally silent) parsing failures; and the Zsh completion functions in
particular parse completions as "name:description" pairs, and thus need
colons in the name to be escaped.

We fix the colon handling of the Zsh completion script by providing
a fixed serialization handler for Zsh.  We avoid all of the other
aforementioned issues by not returning any service names containing
ASCII control characters as completion items.  We also warn the user
upon importing or configuring such a service that its service name will
not be available for shell completion.

Finally, we document this new warning in the manpage and the completion
behavior in the changelog, and add tests for (the Python side of) the
completion machinery, dependent on the current serialization format.

References:
[fish-shell#10874](https://github.com/fish-shell/fish-shell/issues/10874),
[fish-shell#9693](https://github.com/fish-shell/fish-shell/issues/9693),
[fish-shell#751](https://github.com/fish-shell/fish-shell/issues/751),
[fish-shell#10651](https://github.com/fish-shell/fish-shell/issues/10651),
[fish-shell#9847](https://github.com/fish-shell/fish-shell/issues/9847),
[click#2703](https://github.com/pallets/click/issues/2703).