Fix phrasing of "Cannot update %s settings" message
The "Cannot update the global/service settings without actual
settings" messages contained "global" and "service" as (English) format
string parameters, and the actual translatable string contained no
indication what this placeholder actually meant.  This is of course
a horrible way to request a translation for those two messages.  We now
include translatable strings for each of the two possible values of the
metavar ("global settings" and "service-specific settings"), and proper
translation advice for the metavar value and the original translatable
strings.

The manpages have been updated as well, to reflect the slight change in
phrasing of the message.