Update ruff to v0.8.x, reformat and re-lint (part 1 of 2)
We update `ruff` to v0.8.x and reformat the sources.  Most
conspicuously, `assert` statements typically no longer break the
condition to test, and rather push the error message to the next line
instead.  Also, some implicit string concatenations are undone if they
*do* fit onto a single line.

We also re-lint the sources, and commit part of the linting fixes.
Specifically, omissions in the docstring (raised exceptions, return
values) are now marked as linting errors on the docstring itself,
instead of at the omitted object (`raise` or `return` statement).
Alternatively, we add the missing info to the docstring.  We also commit
linting fixes regarding improper usage of underscores in variable names
(leading underscore instead of trailing underscore on temporary
variables).  Finally, we correct some improper calls to the logging
system (root logger instead of application logger).