Incorporate revised wording from the manpage into the README
The revised wording spells out the strength guarantee of the derivation
scheme and the purpose of the configuration file more explicitly.

We intentionally keep the introductory sentences (one-line description
and first sentence of the first explanatory paragraph) different,
because they have different contexts to work with; in particular, the
manpage contains a synopsis section and can refer to the SERVICE metavar
in its one-line description, whereas the README cannot, and should also
mention SSH keys early.