Move manpage directory out of docs, to top-level
The `docs` directory contains sources for the documentation website, to
be uploaded to the WWW.  While also "online" documentation (in the sense
that they are browsable on the customer's computer system), manpages are
not typically distributed online (in the WWW sense).

Besides, the `docs` directory already contains a rendering of the
manpage.