Harmonize the interface for both vault config export handlers
Harmonize the interface for the `storeroom` and the `vault_native`
config export handlers.  Specifically, introduce a common interface
`exporter.ExportVaultConfigDataFunction` which all such handlers must
adhere to, as well as a registry of handlers and a decorator for
registering handlers immediately upon definition.  A new top-level
function `exporter.export_vault_config_data` adhering to this interface
will dispatch to the correct implementation, based on the handler
registry; its use is recommended over calling the handlers directly.

Though it is not encoded in the programmatic interface definition, the
intent is to use common exceptions and exception messages for common
failure modes.  (The intended meanings of the exceptions are listed in
the docstring of `exporter.export_vault_config_data`.)  Specifically,
introduce a new error message constant for when a given handler name is
not registered in the registry, and an exception class
`exporter.NotAVaultConfigError` for when the target path cannot be
decrypted and decoded in the requested vault config encoding format with
the given or inferred storage master key.  Also update some outdated or
wrong aspects of the handler docstrings.

Additionally, the registry is empty upon definition to avoild circular
imports.  A separate call to the
`exporter.find_vault_config_data_handlers` function is necessary to
populate the standard set of handlers.
`exporter.export_vault_config_data` will call this automatically.

Because of this updated interface, the
`vault_native.export_vault_native_data` handler no longer supports
specifying multiple formats to try.  The caller in
`derivepassphrase.cli` already handles trying multiple formats natively,
but additionally no longer dispatches to the handler functions itself,
relying on `exporter.export_vault_config_data` for this task instead.

Finally, update the tests to include the handler registry, the
decorator, the change of parameter names and the change in handling
attempting multiple formats.