Generate debug translations automatically, in code and on the command-line
Introduce a `DebugTranslations` object that returns the enum name of the
given message as its translation, including the parameters to be
interpolated.

Rename the `_format_pot_file` to `_format_po_file` and expand it to emit
either a template file or a debug translation `.po` file.  This differs
slightly in the header, and in whether the translations are empty by
default or filled in with the enum name.  The ad-hoc command-line
interface now accepts options to select the template or the debug
translation, and a way to explicitly override the declared version of
the `.po` template.

Because of negative experience with the `poedit` translations editor,
the message-ID (enum name) is no longer embedded as a (pseudo-)location
of the message string, but rather embedded as a translators' comment.