Properly support trimmed filenames in translation strings
The debug translations object and the `.po` file writer now properly
support translation strings with the filename portion trimmed.
Effectively, these are separate translation strings, and as such they
need separate cache entries and separate `.po` entries.

To fully support this, the translatable string itself now knows how to
trim the filename replacement field, which the debug translations object
now honors.  The `.po` writer still needs the enum name to construct the
message ID, but now it takes an optional transformed string to use
instead of the enum value, if given.

One additional minor fix to the debug translator: some of the emitted
translated messages interpolate other translated messages (such as
metavars).  Thus when interpolating arguments, if the argument is
a translated string, then stringify it before interpolation.  This
avoids printing the `repr` of the inner translated string, which is
otherwise very common for debug translation output.