Fix debug output of messages with trimmed filenames
We do need to keep track of the trimming status when adding those
variants to the cache, because the debug output formatter will discover
the trimmed fields in the template anyway and otherwise attempt to emit
them.  So store the set of trimmed fields in the cache for each message,
and pass it on to the formatter.  We will assume that if a field gets
trimmed, then it had a `None` value, which is true for all the fields we
currently support trimming.