Normalize and unwrap translatable strings (messages and comments)
Rewrite one-paragraph comments and strings as one-line strings, with
implicit line continuation if necessary.  This is how they would appear
anyway if printed via `repr`, and in the `mkdocstrings`-extracted API
documentation.  Further trim the "TRANSLATORS:" prefix from comments; it
is added automatically during `.po` file writing as necessary.

Some minor changes to the docstrings have also been made:

  * `Label`
      - `VAULT_METAVAR_SERVICE`: Trim `as "service_metavar"` from the`
        comments for consistency.
  * DebugMsgTemplate
      - `VAULT_NATIVE_EVP_BYTESTOKEY_RESULT`: Use the correct message.
      - `VAULT_NATIVE_CHECKING_MAC_DETAILS`,
        `VAULT_NATIVE_V02_PAYLOAD_MAC_POSTPROCESSING`: Fix reference to
        `InfoMsgTemplate.VAULT_NATIVE_PARSING_IV_PAYLOAD_AND_MAC`
  * InfoMsgTemplate
      - `VAULT_NATIVE_PARSING_IV_PAYLOAD_MAC`: Use "these terms" instead
        of "this term".

The diff is hard to read in normal "line" mode, but much clearer if read
in "word-diff" mode (with quotes also treated as whitespace).