Fix usage of `mypy --python-version=3.9` and friends
On the one hand, use a small stub file for `tomli`, because `mypy` is
run from all sorts of `hatch` environments or Python virtual
environments that are not set up correctly with `tomli`.

On the other hand, type checking cannot rely on
`try`-`except`-`ImportError` blocks working for type inference purposes,
so variables imported as such need an explicit declaration.  (In this
specific case, this could be alleviated with explicit Python version
checks, but I find that decidedly unpythonic.)