Remove license text from some docstrings in `derivepassphrase.cli`
The license of some parts of the code is clearly an implementation
detail: you have to read the source code to make any use of this license
information.  Accordingly, it does not belong in the rendered API
documentation.  So take it out of the docstrings, and put it into inline
code comments.

Apart from that, also include only *one* copy of the license text and
have all other affected methods refer to that copy instead.