Use `pathlib` for path or filename operations
The interface was already supported by all the user-facing code
(interestingly enough) and by almost all of the internal code.  Using
`pathlib.Path` objects directly instead of manually dealing with string
or byte string collections does cut down on a lot of the otherwise
explicit bookkeeping and context management.

Some changes to the tests are necessary because other (or more)
functions need to be mocked, or other ways of testing successful or
ensuring unsuccessful operations are needed.