Add links and anchors for version/upgrade information and changelog sections
Add separate documentation pages for Python version information and
`derivepassphrase` upgrade notes.  Apart from the informative aspect, it
also serves as an anchor which all code upgrade tasks and TODOs that
belong to the same upgrade step can link to, and which can then be
grepped for in the code.

Relatedly, give the changelog proper, systematic, non-auto-generated
anchor targets.