Add tests API documentation to "dev setup" MkDocs build
Most API documentation is highly deficient at this point, but exposing
it in this way hopefully makes it more likely I'll get to documenting
the tests.  The hope for the future is that it might be faster to
consult the API documentation than to skim through the test file to see
if and where a certain aspect is being tested, whether the classes are
too large, etc.