Rearrange hypothesis tests in the `ssh_agent` module
Arrange the hypothesis tests so that they lie next to the respective
non-hypothesis test, if any, instead of bundling them all in the same
testing class.  By doing this, related tests are closer to each other,
if possible.  Use the hypothesis mark instead of the testing class if
filtering is required.

Additionally, since the helper functions `as_ssh_string`,
`canonicalize1` and `canonicalize2` now are visible in and assigned to
a broader context, give them proper docstrings.