Adjust hypothesis deadlines globally and redo the profiles and slowdown factors
Previously, we would define a "worst-case" hypothesis profile for use
under instrumentation (coverage measurement), and decorate every
hypothesis test which we experienced to have deadline problems with this
decorator.  This step was manual, and very error-prone when generating
new hypothesis tests.  Conceptually, it also indicated that adjusted
deadlines applied for some tests, but not for others, even though the
deadline adjustment is actually a side-effect of global instrumentation.

We therefore now adopt a global solution, where the pytest configuration
file detects which coverage instrumentation (if any) is running, and
adjusts the default hypothesis settings appropriately.  The decorator is
removed.

In the course of this, we re-evaluated the profiles and the slowdown
factors, instead of assuming the worst-case instrumentation.  hypothesis
also defines a "ci" profile for use in continuous integration, together
with some commentary on how this profile is expected to be used.  We
cede this profile name to hypothesis (and remove it from our testing
matrix in hatch), and define our own "intense" profile for high-volume
hypothesis testing based on the "ci" profile, to be invoked manually as
a "pre-release" (or perhaps "nightly") run.

Finally, because the test suite now is rather slow in sequential use, we
add `--dist=worksteal` to the test options to better distribute load
during parallel testing.  We choose this over other more "balanced"
strategies because our single-test runtimes wildly fluctuate thanks to
hypothesis, so work stealing appears to be the most reliable
distribution heuristic for us.