Convert Parametrizations enums to Parametrize namespace
Switch out `enum.Enum` for `types.SimpleNamespace` to collect the
various `pytest.mark.parametrize` decorators: it works equally well in
practice when coupled with a type checker, and it avoids having to
unpack the enum value when actually calling the decorator.  Opt for the
shorter name `Parametrize` instead of `Parametrizations`.

Also consolidate the two versions of the `VAULT_CONFIG_FORMATS_DATA`
parametrization set by moving it into a dedicated namespace in the
top-level `tests` package.