Implement all parametrization-related changes suggested in the TODOs
Most of these are straight-forward cases of renaming, consolidating, or
yielding additional parameters.

Regarding the TODOs in the `ssh_agent` test module dealing with `skip`
and `xfail` messages, I decided to follow [Paul Ganssle's use of
`xfail`][GANSSLE_XFAIL]: "use `skip` for tests that *aren't supposed to
work*", and "use `xfail` for tests that *should, but currently don't
pass*".  (In other words, for `skip`, the failure is external to the
system under test and cannot be accomodated for; for `xfail`, it is
internal, or part of the system under test's job to accomodate for it.)
For the aforementioned tests for the `ssh_agent` module, both `skip`
messages occur because of unsatisfied assumptions on the running SSH
agent (a test key isn't loaded, or one kind of failure mode doesn't
occur with this agent), so they shall both remain `skip` messages.

[GANSSLE_XFAIL]: https://blog.ganssle.io/articles/2021/11/pytest-xfail.html