Add xfailing tests for indentation and line breaks in vault configurations
Though the on-disk vault configurations are not intended to be managed
by the user, it is valuable for debugging and lay examination for them
to be readable and editable anyway; doubly so if the configuration is
being exported.  So ensure that when writing out the configuration, we
add indents and linebreaks.

This commit only adds the xfailing tests for this functionality, not the
implementation itself.  Additionally, since there was no simple test
akin to config importing but for config exporting, we added the latter
on the basis of the former, and renumbered the other export (failure)
tests accordingly.