Warn the user when `derivepassphrase vault --notes` is ineffective
In 7b0f4e121a5e688b59abad73b8b60bacfc7d02ed, we adjusted
`derivepassphrase` to only act upon `--notes` if `--config` is also
given, for compatibility with vault(1).  However, this is incompatible
with previous versions of `derivepassphrase`, and potentially confusing
even to vault(1) users.  As such, `derivepassphrase vault` now emits
a warning if it is called with `--notes` but without `--config`.