Reintroduce a "modern" editor interface à la git-commit or git-rebase
Reintroduce the previous git-like editor interface as the "modern"
editor interface to the "vault" subcommand, and provide explicit options
`--modern-editor-interface` and `--vault-legacy-editor-interface`,
defaulting to the latter for compatibility.

Furthermore, for the legacy interface, introduce a backup copy of the
notes to guard against accidental data loss, because the legacy
interface offers no way to abort mid-editing, making fatal mistakes all
the more likely.