Document the modern editor interface
Update the manpages to include the corrected and extended handling of
the `--notes` config option.  That is, document the
`--modern-editor-interface` and `--vault-legacy-editor-interface`
options, and the new warning messages about `--notes` requiring
`--config` and that a backup copy of the notes have been made.

Some minor adjustments to the vault(1) legacy editor interface warning
messages were made to ease inclusion in the manpage while making the
intent clear.  Additionally, derivepassphrase with the legacy interface
correctly fills in the old notes contents if any, instead of
unconditionally filling in the placeholder text.