Prepare `bump-my-version` for this project
Prepare and configure the use of `bump-my-version` to automate keeping
version numbers synchronized everywhere.  From a cursory search, this
appears to be the most mature and actively developed tool of this kind
with support for `pyproject.toml` and multiple files with file-specific
replacements.  Some minor warts remain (non-recursive globbing,
duplication of the version number, incorrect calculation of the next
release due to missing true support for pre-release segments), but they
can be worked around, for now.