Add quality control scripts
Add a quality control script `qc_auto.py` that calls the linter, the
formatter, the type checker, the test suite and the documentation
builder, depending on what branch we are on.  It is intended to be
usable as a pre-commit hook.

Add another quality control script `man_diagnostics.py` that checks that
the diagnostics documented in the manpages are complete, and match the
enum values within `derivepassphrase`.  It relies on annotations
(comments) in the manpages to map description texts to the respective
enum values (the mapping is not one-to-one).  It also does
a (rudimentary) check that every warning and error message is mentioned
somewhere in the source tree (besides the messages module that defines
those messages).