Fix automatic quality control handling of too-long lines
In general, we want the formatter to deal with too-long lines (E501) as
much as possible by itself, and the linter to warn us on E501 lines the
formatter *cannot* deal with.  However, just naively running the linter
and formatter causes the linter to abort on any E501 lines it
encounters, because the linter cannot auto-fix them.  So we want to run
the linter twice, once before the formatter and once after the
formatter, while ignoring E501 violations in the first run.  (But
because E501 noqa markers are now treated as unnecessary noqa markers
during the first linting run, we actually need to ignore both E501 and
RUF100 in the first run.)