Automatically check and build the translations on master via hatch-gettext
Include the configuration for `hatch-gettext` to automatically build the
translations during build time, and include them in the wheel.  Further
support a quality control check that tests the PO template for
reproducability.  The PO template generator now supports passing
a specific build timestamp, and realistically, the reproducability check
needs this timestamp to be specified (via `SOURCE_DATE_EPOCH`).

Actually *using* these translations directly from the installed wheel is
not yet supported, because it is actually somewhat difficult to use the
`gettext` MO-finding machinery here directly: `gettext.find` requires
the directories it searches to be on-disk directories, whereas the wheel
installer does not guarantee that installed packages are present as
files on the system (instead of, say, as a ZIP archive).  The sanest way
forward seems to be reimplementing `gettext.find` with support for
`importlib.metadata.as_file`, and then constructing
`gettext.GNUTranslations` objects directly.  This would also be the
correct time to introduce a new enum for all these automatically
calculated platform-specific special locale directories.  (TODO.)